refactor add-prompt controller to generate-llm-response endpoint
