--- src/controllers/chat-messages/index.ts
@@ -1,1 +1,0 @@
-export * from './create';
