+++ src/controllers/chats/index.ts
@@ -1,4 +1,4 @@
 export * from './list';
 export * from './create';
 export * from './show';
-export * from './add-prompt';
+export * from './generate-llm-response';
