+++ src/routes/v1/chats.ts
@@ -1,12 +1,17 @@
 import { Router } from 'express';
 
-import { list, show, create, addPrompt } from '../../controllers/chats';
+import {
+  list,
+  show,
+  create,
+  generateLLMResponse
+} from '../../controllers/chats';
 
 const router = Router();
 
 router.get('/', list);
 router.get('/:id', show);
 router.post('/', create);
-router.patch('/:id/prompt', addPrompt);
+router.patch('/:id/generate-llm-response', generateLLMResponse);
 
 export default router;
