+++ src/services/index.ts
@@ -1,1 +1,3 @@
 export * from './chat';
+export * from './chat-message';
+export * from './llm';
